// src/lib/types/youtube.ts
// Generated by https://quicktype.io

export interface YoutubeSnippet {
	kind: YoutubeSnippetKind;
	etag: string;
	id: ID;
	snippet: Snippet;
}

export interface ID {
	kind: IDKind;
	videoId: string;
}

export enum IDKind {
	YoutubeVideo = 'youtube#video'
}

export enum YoutubeSnippetKind {
	YoutubeSearchResult = 'youtube#searchResult'
}

export interface Snippet {
	publishedAt: string;
	channelId: ChannelID;
	title: string;
	description: string;
	thumbnails: Thumbnails;
	channelTitle: ChannelTitle;
	liveBroadcastContent: LiveBroadcastContent;
	publishTime: string;
}

export enum ChannelID {
	UClnCId37Ib0QSFxCqnXnbvQ = 'UClnCId37ib0qSFxCqnXnbvQ'
}

export enum ChannelTitle {
	TheRidgeTechTeam = 'The Ridge Tech Team'
}

export enum LiveBroadcastContent {
	None = 'none'
}

export interface Thumbnails {
	default: Default;
	medium: Default;
	high: Default;
}

export interface Default {
	url: string;
	width: number;
	height: number;
}

// Sermon after manipulation
export interface Sermon {
	id: string;
	info: Info;
	description: string;
	thumbnail: Thumbnail;
}

export interface Info {
	date: DateClass;
	title: string;
}

export interface DateClass {
	day: string;
	month: string;
	year: string;
}

export interface Thumbnail {
	url: string;
	width: number;
	height: number;
}